fixed some bugs.

+++ Apartamentsapp/Flat.cs
@@ -9,8 +9,8 @@ namespace Apartamentsapp
     public class Flat
     {
         public string FlatName { get; set; }
-        public double ApartmentPrice { get; set; }
-        public double Area { get; set; }
+        public decimal ApartmentPrice { get; set; }
+        public decimal Area { get; set; }
         public District District { get; set; }
     }
 }
